Move 2SCM code to -typing feature.
(Step 1 from #7444)
