+++ packages/ckeditor5-typing/tests/manual/1301/1.md
@@ -1,4 +1,4 @@
-## Two-steps caret movement [#1301](https://github.com/ckeditor/ckeditor5-engine/issues/1301)
+## Two-steps caret movement [ckeditor5-engine#1301](https://github.com/ckeditor/ckeditor5-engine/issues/1301)
 
 1. Put the selection at the end of the content.
 2. Press the <kbd>←</kbd> key 3 times.
