+++ packages/ckeditor5-typing/tests/manual/two-step-caret.html
